requirements.txt
git+https://github.com/facebookresearch/ImageBind.git
ipython
pandas
Pillow
requests
seaborn
torch
